refactor(probe): complete ndp migration with services and rewrite hooks
